Feat: 추가 바우처 ID 조회, 삭제시 존재여부 확인

+++ src/main/java/org/prgrms/kdt/engine/voucher/controller/VoucherRestController.java
@@ -3,9 +3,12 @@ package org.prgrms.kdt.engine.voucher.controller;
 import org.prgrms.kdt.engine.voucher.domain.Voucher;
 import org.prgrms.kdt.engine.voucher.dto.VoucherDto;
 import org.prgrms.kdt.engine.voucher.service.VoucherService;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Map;
+import java.util.Optional;
 import java.util.UUID;
 
 @RestController
@@ -31,11 +34,21 @@ public class VoucherRestController {
 
     @GetMapping("/api/v1/vouchers/{voucherId}")
     public Voucher findVoucher(@PathVariable("voucherId") UUID voucherId) {
-        return voucherService.getVoucher(voucherId);
+        Optional<Voucher> voucher = voucherService.getVoucher(voucherId);
+
+        if (voucher.isEmpty())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "voucher does not exist");
+
+        return voucher.get();
     }
 
     @DeleteMapping("/api/v1/vouchers/{voucherId}")
     public String deleteVoucher(@PathVariable("voucherId") UUID voucherId) {
+        Optional<Voucher> voucher = voucherService.getVoucher(voucherId);
+
+        if (voucher.isEmpty())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "voucher does not exist");
+
         voucherService.deleteVoucher(voucherId);
         return "deleted voucher " + voucherId;
     }
